+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features
 + Reduce width of select column in tables by removing padding and replacing title with icon
 + Move ticket status before title and hide its label to tooltip
-+ Improve titles in account section of menu
++ Improve payments title in menu
 
 ### Fixes